pyproject.toml
[project]
name = "web-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "curl-cffi>=0.14.0",
    "nodriver>=0.48.1",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "pyyaml>=6.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/web_scraping", "src/storage"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""
